extract template helpers
