+++ lib/templates/index.js
@@ -1,3 +1,8 @@
 'use strict';
 
-module.exports.table = require('./table.ejs');
+exports.table = require('./table.ejs');
+
+exports.projectHeading = require('./projectHeading');
+exports.projectHours = require('./projectHours');
+exports.hourCellClass = require('./hourCellClass');
+exports.rowStyle = require('./rowStyle');
