+++ lib/templates/projectHeading.js
@@ -1,0 +1,24 @@
+'use strict';
+
+module.exports = (projects, leaveTypes, projectKey) => {
+
+    if (projectKey.startsWith('LeaveType-')) {
+        const leaveId = +projectKey.substr(10);
+        const leave = leaveTypes.find((p) => p.id === leaveId);
+
+        if (leave) {
+            return leave.name;
+        }
+    }
+
+    if (projectKey.startsWith('Project-')) {
+        const projectId = +projectKey.substr(8);
+        const project = projects.find((p) => p.id === projectId);
+
+        if (project) {
+            return project.name;
+        }
+    }
+
+    return projectKey;
+};
