+++ lib/templates/projectHours.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = (projectData, rowHours) => {
+
+    if (projectData && projectData.hours) {
+        return projectData.hours;
+    }
+
+    if (projectData && projectData.scheduled && rowHours <= 0) {
+        return projectData.scheduled;
+    }
+
+    return '';
+};
