+++ lib/templates/table.ejs
@@ -1,4 +1,5 @@
 <% const Utils = require('../utils'); -%>
+<% const Templates = require('../templates'); %>
 
 <table id="results-3048m" style="clear: both;" class="widget-wrapper">
 
@@ -6,7 +7,7 @@
         <tr>
             <th>Date</th>
             <% Object.keys(timeEntries.totals).forEach((k) => { %>
-            <th colspan="2" style="padding-right: 14px;"><%= Utils.formatProjectHeading(projects, leaveTypes, k) %></th>
+            <th colspan="2" style="padding-right: 14px;"><%= Templates.projectHeading(projects, leaveTypes, k) %></th>
             <% }); %>
         </tr>
     </thead>
@@ -18,13 +19,13 @@
         %>
 
         <% if (dayData.display) { %>
-        <tr class="tk-time-tracker-row" style="<%= Utils.getRowStyle(dayData) %>">
+        <tr class="tk-time-tracker-row" style="<%= Templates.rowStyle(dayData) %>">
 
             <td style="white-space: nowrap;padding-right: 14px;"><%= dateKey %></td>
 
             <% Object.keys(timeEntries.totals).forEach((projectKey) => { const projectData = dayData.projects[projectKey]; %>
-            <td class="<%- Utils.getHourCellClass(projects, leaveTypes, projectKey, projectData) %>"><div class="tk-hours">
-                <%- Utils.getProjectHours(projectData, rowHours) %>
+            <td class="<%- Templates.hourCellClass(projects, leaveTypes, projectKey, projectData) %>"><div class="tk-hours">
+                <%- Templates.projectHours(projectData, rowHours) %>
             </div></td>
             <td style="padding-right: 14px;">
                 <%- projectData && projectData.notes || '' %>
